+++ src/Contracts/RpcResultDto.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Uepg\LaravelSybase\Contracts;
+
+interface RpcResultDto
+{
+    public static function fromArray(array $row): static;
+}